backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.36
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-multipart>=0.0.20
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
aiofiles>=24.1.0
qrcode>=8.0
pyotp>=2.9.0
MetaTrader5>=5.0.45
